requirements.txt
aiogram==3.18.0
aiohttp==3.11.13
python-dotenv==1.0.1